back-end/usuarios/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.1.0
aiokafka==0.12.0
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.2
minio==7.2.12
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-prometheus==0.50b0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
redis[hiredis]==5.2.1
pytest==8.3.4
pytest-asyncio==0.24.0
